Added coloring to the rank panel
